Displayed stats improvement
- Fire Delay is now replaced by Firerate
- level is also shown

+++ src/model/towers/ElGatoModel.java
@@ -39,7 +39,7 @@ public class ElGatoModel extends TowerModel{
     }
 
     @Override
-    public int getFireRate() {
+    public double getFireRate() {
         return attackSpeed;
     }
 }
